Add files via upload
This file is used on the ghpcc to break the yearly shapefile into many smaller shapefiles, one for each time step.  It then loads in each time step shapefile and computes the average probability of a lethal ship strike for each grid cell and saves to an .rda file that can be pulled off the clusters and mapped on a local computer.